requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=requirements-dev.txt --strip-extras requirements-base.in
#
asgiref==3.8.1
    # via django
django==4.1.13
    # via
    #   -r requirements-base.in
    #   django-ajax-selects
    #   django-crispy-forms
    #   django-filter
    #   djangorestframework
django-ajax-selects==3.0.3
    # via -r requirements-base.in
django-crispy-forms==2.0
    # via -r requirements-base.in
django-filter==23.5
    # via -r requirements-base.in
djangorestframework==3.15.1
    # via -r requirements-base.in
djangorestframework-api-key==3.0.0
    # via -r requirements-base.in
packaging==24.2
    # via djangorestframework-api-key
pillow==11.0.0
    # via -r requirements-base.in
sqlparse==0.5.3
    # via django
